requirements.txt
Flask==3.1.0
langchain==0.3.23
langchain_community==0.3.21
langchain_core==0.3.54
langchain_text_splitters==0.3.8
pytesseract==0.3.13
python-dotenv==1.1.0
Werkzeug==3.1.3
